refactor!: moves device.py logic to gemini subfolder
